// src/clients/HubspotEmailClient/constants.ts
export enum HubspotEmails {
  INVITE_CUSTOMER = 'INVITE_CUSTOMER',
  INVITE_SUPPLIER = 'INVITE_SUPPLIER',
  CUSTOMER_INVITATION_APPROVED = 'CUSTOMER_INVITATION_APPROVED',
  SUPPLIER_INVITATION_APPROVED = 'SUPPLIER_INVITATION_APPROVED',
  CUSTOMER_INVITATION_DECLINED = 'CUSTOMER_INVITATION_DECLINED',
  SUPPLIER_INVITATION_DECLINED = 'SUPPLIER_INVITATION_DECLINED',
  EMISSION_ALLOCATION_REQUESTED = 'EMISSION_ALLOCATION_REQUESTED',
  EMISSION_ALLOCATION_SUBMITTED = 'EMISSION_ALLOCATION_SUBMITTED',
  EMISSION_ALLOCATION_ACCEPTED = 'EMISSION_ALLOCATION_ACCEPTED',
  EMISSION_ALLOCATION_REJECTED = 'EMISSION_ALLOCATION_REJECTED',
  EMISSION_ALLOCATION_UPDATED = 'EMISSION_ALLOCATION_UPDATED',
  EMISSION_ALLOCATION_DELETED = 'EMISSION_ALLOCATION_DELETED',
  UNABLE_TO_INVITE_COMPANY = 'UNABLE_TO_INVITE_COMPANY',
  REGISTRATION_UNSUCCESSFUL = 'REGISTRATION_UNSUCCESSFUL',
  JOINING_INVITATION_DECLINED = 'JOINING_INVITATION_DECLINED',
  RESEND_INVITE_TO_JOIN = 'RESEND_INVITE_TO_JOIN',
  DATA_SHARE_REQUEST = 'DATA_SHARE_REQUEST',
}